% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binarySearchC}
\alias{binarySearchC}
\title{bisection}
\usage{
binarySearchC(x, target)
}
\arguments{
\item{x}{target sequence}

\item{target}{the target number}
}
\value{
mid target value subscript
}
\description{
Find the target value using the dichotomy method
}
\examples{
\dontrun{
  x <- seq(1:50)
  target <- 8
  result <- binarySearchC(x, target)
  print(result)
}
}
