% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA22231050.R
\name{bisection}
\alias{bisection}
\title{bisection}
\usage{
bisection(x, target)
}
\arguments{
\item{x}{target sequence}

\item{target}{the target number}
}
\value{
mid target value subscript
}
\description{
Find the target value using the dichotomy method
}
\examples{
\dontrun{
x <- seq(1:100)
bisection(x, 15)
}
}
