% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA22231050.R
\name{newton}
\alias{newton}
\title{newton}
\usage{
newton(f, f1, x0, tol, max_iter = 1000)
}
\arguments{
\item{f}{equation}

\item{f1}{Derivatives of Eq.}

\item{x0}{the init number}

\item{tol}{the tolerant number}

\item{max_iter}{the max number of loop}
}
\value{
x solution
}
\description{
Newton's method to find roots
}
\examples{
\dontrun{
ans <- newton(f = function(x) x^2 - 4, f1 = function(x) 2*x, x0 = -5, tol = 1e-6)
}
}
