% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newtonMethodC}
\alias{newtonMethodC}
\title{newton}
\usage{
newtonMethodC(x0, tol, max_iter)
}
\arguments{
\item{x0}{the init number}

\item{tol}{the tolerant number}

\item{max_iter}{the max number of loop}
}
\value{
mid target value subscript
}
\description{
Newton's method to find roots
}
\examples{
\dontrun{
x0 <- -9
tol <- 1e-6
max_iter <- 100
result <- newtonMethodC(x0, tol, max_iter)
print(result)
}
}
